requirements.txt
numpy==1.26.0
pandas==2.2.2
pynrrd==1.0.0
nibabel==5.2.1
tqdm==4.66.4
scikit-image==0.24.0
scikit-learn==1.5.1
torch==2.3.1
torchvision==0.18.1
matplotlib==3.9.2
opencv-python==4.10.0.84
plotly==5.22.0
openpyxl==3.1.5
kaleido==0.2.1
